Rename client_connection package name to Go naming conventions
